Temp for cook food thermometer placed, better than previously
The dot at the thermometer (red at first, then yellow, then green) shows what stage the food is in concerning cook times. If green dot, time has passed for the food to be completely safe, yellow only indicates that food is being cooked, and red is when the thermometer is off the food, or stove is off.